requirements.txt
# Django и основные зависимости
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# База данных
psycopg2-binary==2.9.9

# Кеширование и очереди
redis==5.0.1
celery==5.3.4
django-celery-beat==2.5.0

# WebSocket для уведомлений
channels==4.0.0
channels-redis==4.1.0

# Аутентификация и безопасность
django-allauth==0.57.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7

# Медиа файлы и статика
Pillow==10.1.0
whitenoise==6.6.0

# Разработка
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Production
gunicorn==21.2.0
python-decouple==3.8
